requirements.txt
flask
boto3
datetime
mysql-connector-python